[Feature] Setup HomeView model, add portfolia list, add columnTitles to AllCoinList, PortfolioList
