+++ SwiftfulCrypto/Core/Home/ViewModels/HomeViewModel.swift
@@ -1,0 +1,20 @@
+//
+//  HomeViewModel.swift
+//  SwiftfulCrypto
+//
+//  Created by Pasya Sahas on 10/29/24.
+//
+
+import Foundation
+
+class HomeViewModel: ObservableObject {
+    @Published var allCoins: [CoinModel] = []
+    @Published var portfolioCoins: [CoinModel] = []
+    
+    init(){
+        DispatchQueue.main.asyncAfter(deadline: .now() + 2.0) {
+            self.allCoins.append(DeveloperPreview.instance.coin)
+            self.portfolioCoins.append(DeveloperPreview.instance.coin)
+        }
+    }
+}
